project-backend/requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
aws-cdk.aws-amplify-alpha
argparse
boto3